feat: 抽离表格
